examples/prompted_pipeline/pyproject.toml
[project]
name = "prompted_pipeline"
version = "0.1.0"
description = "Multi-step pipeline example using a simple prompt loader"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "restack-ai>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
